+++ assets/javascripts/kitten/components/icons/clock-icon.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export const StudingIcon = props => (
+export const ClockIcon = props => (
   <svg xmlns="http://www.w3.org/2000/svg"
        viewBox="0 0 11 11"
        width="10px"