+++ spec/dummy/client/entries/app-kitten.js
@@ -50,13 +50,14 @@ import { KarlSideLayout } from 'kitten/karl/layout/side-layout'
 // Icons
 import GrabberIcon from 'kitten/components/icons/grabber-icon'
 import { ArrowRightIcon } from 'kitten/components/icons/arrow-right-icon'
-import { CbIcon } from 'kitten/components/icons/cb-icon'import { MenuIcon } from 'kitten/components/icons/menu-icon'
+import { CbIcon } from 'kitten/components/icons/cb-icon'
+import { MenuIcon } from 'kitten/components/icons/menu-icon'
 import { HeartIcon } from 'kitten/components/icons/heart-icon'
 import { InformationIcon } from 'kitten/components/icons/information-icon'
 import { KissKissBankBankLogoIcon } from 'kitten/components/icons/kisskissbankbanklogo-icon'
 import { FacebookIcon } from 'kitten/components/icons/facebook-icon'
 import { LinkedinIcon } from 'kitten/components/icons/linkedin-icon'
-import { StudingIcon } from 'kitten/components/icons/studing-icon'
+import { ClockIcon } from 'kitten/components/icons/clock-icon'
 import { LocationIcon } from 'kitten/components/icons/location-icon'
 import { MasterCardIcon } from 'kitten/components/icons/mastercard-icon'
 import { TwitterIcon } from 'kitten/components/icons/twitter-icon'
@@ -238,7 +239,7 @@ ReactOnRails.register({
   HeartIcon,
   FacebookIcon,
   LinkedinIcon,
-  StudingIcon,
+  ClockIcon,
   LocationIcon,
   MasterCardIcon,
   TwitterIcon,